fastapi-vnstock-server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
vnstock
anthropic
redis
httpx
psycopg[binary]
